Added NMDA_Leak mechanism to GranCellSol10_mod. Disabled TABLE in KCa mod file.
Here I'm trying to get the neuroConstruct-with-mod-files model to match
the behaviour of the original NEURON+Python cell. So, I started by importing
the NMDA_Leak mechanism in the nC project and adding it to GranCellSol10_mod.
Then I noticed that the model defined in GRANULE_Cell.py sets
h.usetable_GRANULE_KCA = 0
and I thought of emulating this by commenting out the TABLE and DEPEND lines
in the GRANULE_KCA.mod file for the neuroConstruct project (I guess another
way of doing this would have been to manually add some extra hoc code to
the simulation).
The match is not perfect yet; there must be some other problem.